adding stylesheet for Timeline Tools panel for scheudle changes
