+++ src/app/containers/LongTermSchedule/LongTermSchedule.css
@@ -18,3 +18,7 @@
 .scheduling-timeline .shift-card:hover {
     margin: 1px;
 }
+
+.scheduling-timeline .react-calendar-timeline .rct-header .rct-label-group {
+    background: #327AB7;
+}
